Extract KAPI PDB reconciliation into dedicated file
for a better overview
